Separate node-only tests from karma tests.
