+++ package.json
@@ -26,22 +26,22 @@
     "crypto-ld": "digitalbazaar/crypto-ld#extract-key-pairs",
     "node-forge": "^0.9.1",
     "semver": "^7.3.2",
-    "sodium-native": "^3.1.1"
+    "sodium-native": "^3.2.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.10.2",
-    "@babel/plugin-transform-modules-commonjs": "^7.10.1",
-    "@babel/plugin-transform-runtime": "^7.10.1",
-    "@babel/preset-env": "^7.10.2",
-    "@babel/runtime": "^7.10.2",
+    "@babel/core": "^7.10.5",
+    "@babel/plugin-transform-modules-commonjs": "^7.10.4",
+    "@babel/plugin-transform-runtime": "^7.10.5",
+    "@babel/preset-env": "^7.10.4",
+    "@babel/runtime": "^7.10.5",
     "babel-loader": "^8.1.0",
     "chai": "^4.2.0",
     "chai-bytes": "^0.1.2",
     "cross-env": "^7.0.2",
-    "eslint": "^7.2.0",
+    "eslint": "^7.5.0",
     "eslint-config-digitalbazaar": "^2.5.0",
-    "eslint-plugin-jsdoc": "^27.0.4",
-    "karma": "^5.0.9",
+    "eslint-plugin-jsdoc": "^30.0.3",
+    "karma": "^5.1.0",
     "karma-babel-preprocessor": "^8.0.1",
     "karma-chai": "^0.1.0",
     "karma-chrome-launcher": "^3.1.0",
@@ -49,12 +49,12 @@
     "karma-mocha-reporter": "^2.2.5",
     "karma-sourcemap-loader": "^0.3.7",
     "karma-webpack": "^4.0.2",
-    "mocha": "^7.2.0",
+    "mocha": "^8.0.1",
     "mocha-lcov-reporter": "^1.3.0",
-    "multibase": "^0.7.0",
-    "multicodec": "^1.0.1",
+    "multibase": "^2.0.0",
+    "multicodec": "^1.0.4",
     "nyc": "^15.1.0",
-    "webpack": "^4.43.0"
+    "webpack": "^4.44.0"
   },
   "nyc": {
     "exclude": [
@@ -86,7 +86,7 @@
   ],
   "scripts": {
     "test": "npm run test-node",
-    "test-node": "cross-env NODE_ENV=test mocha --preserve-symlinks -t 30000 -A -R ${REPORTER:-spec} tests/*.spec.js",
+    "test-node": "cross-env NODE_ENV=test mocha --preserve-symlinks -t 30000 -A -R ${REPORTER:-spec} tests/**/*.spec.js",
     "test-karma": "karma start",
     "benchmark": "node benchmark/benchmark.js",
     "coverage": "cross-env NODE_ENV=test nyc --reporter=lcov --reporter=text-summary npm run test-node",
