+++ tests/unit/sign-verify.spec.js
@@ -3,8 +3,8 @@
  */
 'use strict';
 
-const {Ed25519VerificationKey2018} = require('..');
-const mockKey = require('./mock-key');
+const {Ed25519VerificationKey2018} = require('../../src/Ed25519VerificationKey2018');
+const mockKey = require('../mock-key.json');
 
 const keyPair = new Ed25519VerificationKey2018({
   publicKeyBase58: mockKey.publicKeyBase58,
@@ -15,7 +15,7 @@ const signer = keyPair.signer();
 const verifier = keyPair.verifier();
 
 // the same signature should be generated on every test platform
-// (eg. browser, node8, node12)
+// (eg. browser, node12, node14)
 const targetSignatureBase64 = 'nlQC1bVF6TMN6cAEJllRGK5orHm5+n4Ih46mu' +
   'RYgQhTl8J9SR82fEPq7IEAmT9GprBrcRKJzxUk0Eo+yU92zCg==';
 
@@ -23,6 +23,7 @@ describe('sign and verify', () => {
   it('works properly', async () => {
     const data = Buffer.from('test 1234');
     const signature = await signer.sign({data});
+
     signature.toString('base64').should.equal(targetSignatureBase64);
     const result = await verifier.verify({data, signature});
     result.should.be.true;
